refactor: split VRMError, move pure-glTF loaders to GLTFCore
Phase 1b: completes the originally-planned Phase 1 file extractions
that were deferred because the underlying coupling was structural.

Split VRMError → GLTFError + VRMError:
- GLTFError (in GLTFCore) — 14 glTF-spec cases: invalidGLBFormat,
  invalidJSON, unsupportedVersion, missingBuffer, invalidAccessor,
  missingTexture, invalidImageData, invalidMesh, missingVertexAttribute,
  deviceNotSet, invalidPath, invalidMaterial, unsupportedRequiredExtension,
  loadingCancelled
- VRMError (slim, in VRMMetalKit) — 3 VRM-specific cases:
  missingVRMExtension, missingRequiredBone, invalidMeta

Moved to Sources/GLTFCore/:
- Loader/BufferLoader.swift (731 lines, accessor decoding)
- Loader/BufferPreloader.swift
- Loader/GLTFParser.swift (792 lines, glTF 2.0 JSON parsing + glTF
  document types: GLTFDocument, GLTFNode, GLTFMesh, GLTFMaterial,
  GLTFPrimitive, GLTFTexture, GLTFImage, GLTFSampler, GLTFBuffer,
  GLTFBufferView, GLTFAccessor, etc.)
- Loader/TextureLoader.swift
- Loader/ParallelTextureLoader.swift
- Builder/GLTFDocumentBuilder.swift

56 throw sites rewritten; 1 catch site updated; DocC links retargeted.
BufferLoader.filePath promoted to public for cross-module access.
GLTFNode gained a public memberwise init so AnimationTests can build.

ParallelMaterialLoader, ParallelMeshLoader, VRMExtensionParser remain
in VRMMetalKit — they are the mixed glTF/VRM seams scheduled for
Phase 2 untangling.

Verification: swift build clean; all 1441 VRMMetalKitTests pass.
