+++ Sources/GLTFCore/Builder/GLTFDocumentBuilder.swift
