+++ Sources/GLTFCore/Loader/BufferLoader.swift
