+++ Sources/GLTFCore/Loader/BufferPreloader.swift
