+++ Sources/GLTFCore/Loader/ParallelTextureLoader.swift
