+++ Sources/GLTFCore/Loader/TextureLoader.swift
